Harvester Optimization Update
Everything should be fixed now
Added room.init function that should run once upon room creation and set
memory values that we want to see
